requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
black==20.8b1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
conda==4.9.2
conda-package-handling==1.7.0
cryptography==2.9.2
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.9
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
menuinst==1.4.16
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2020.4
pywin32==227
regex==2020.11.13
requests==2.23.0
ruamel-yaml==0.15.87
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
tqdm==4.46.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
